# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/call-status": {"post": {"summary": "Call Status", "operationId": "call_status_call_status_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/schedule-call": {"post": {"summary": "Schedule Call", "operationId": "schedule_call_schedule_call_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScheduledCall"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/schedule_calls": {"post": {"summary": "Post Schedule Calls", "operationId": "post_schedule_calls_schedule_calls_post", "parameters": [{"required": true, "schema": {"type": "string", "title": "Phone Number"}, "name": "phone_number", "in": "query"}, {"required": true, "schema": {"type": "string", "format": "date-time", "title": "Start Date"}, "name": "start_date", "in": "query"}, {"required": true, "schema": {"type": "string", "format": "date-time", "title": "End Date"}, "name": "end_date", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Morning Time"}, "name": "morning_time", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Afternoon Time"}, "name": "afternoon_time", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Evening Time"}, "name": "evening_time", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Morning Minute"}, "name": "morning_minute", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Afternoon Minute"}, "name": "afternoon_minute", "in": "query"}, {"required": true, "schema": {"type": "integer", "title": "Evening Minute"}, "name": "evening_minute", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/cancel-call": {"post": {"summary": "Cancel Call", "operationId": "cancel_call_cancel_call_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ScheduledCall"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/cancel-calls": {"post": {"summary": "Cancel Schedule Calls", "operationId": "cancel_schedule_calls_cancel_calls_post", "parameters": [{"required": true, "schema": {"type": "string", "title": "Phone Number"}, "name": "phone_number", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/list-scheduled-calls": {"get": {"summary": "List Schedule Calls", "operationId": "list_schedule_calls_list_scheduled_calls_get", "parameters": [{"required": true, "schema": {"type": "string", "title": "Phone Number"}, "name": "phone_number", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ScheduledCall"}, "type": "array", "title": "Response List Schedule Calls List Scheduled Calls Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/add-participant": {"post": {"summary": "Add Participant", "operationId": "add_participant_add_participant_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Participant"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/delete-participant": {"post": {"summary": "Delete Participant", "operationId": "delete_participant_delete_participant_post", "parameters": [{"required": true, "schema": {"type": "string", "title": "Participant Study Id"}, "name": "participant_study_id", "in": "query"}, {"required": true, "schema": {"type": "string", "title": "Phone Number"}, "name": "phone_number", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/call-log": {"get": {"summary": "Api Call Log", "operationId": "api_call_log_api_call_log_get", "parameters": [{"required": true, "schema": {"type": "string", "title": "Call Sid"}, "name": "call_sid", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CallLog"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/call-list": {"get": {"summary": "Api Call List", "operationId": "api_call_list_api_call_list_get", "parameters": [{"required": true, "schema": {"type": "string", "title": "Participant Study Id"}, "name": "participant_study_id", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/CallLogHeader"}, "type": "array", "title": "Response Api Call List Api Call List Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/participant-list": {"get": {"summary": "Api Participant List", "operationId": "api_participant_list_api_participant_list_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/Participant"}, "type": "array", "title": "Response Api Participant List Api Participant List Get"}}}}}}}, "/": {"get": {"summary": "Read Index1", "operationId": "read_index1__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/favicon.ico": {"get": {"summary": "Read Index2", "operationId": "read_index2_favicon_ico_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"CallLog": {"properties": {"call_sid": {"type": "string", "title": "Call Sid"}, "phone_number": {"type": "string", "title": "Phone Number"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}, "participant_study_id": {"type": "string", "title": "Participant Study Id", "default": "unknown"}, "rejected": {"type": "string", "title": "Rejected", "default": "false"}, "previous_rejects": {"type": "integer", "title": "Previous Rejects"}, "history": {"items": {"items": [{"type": "string"}, {"type": "string"}], "type": "array", "maxItems": 2, "minItems": 2}, "type": "array", "title": "History"}, "syntax_grade": {"items": {}, "type": "array", "title": "Syntax Grade", "default": []}, "perplexity_grade": {"type": "number", "title": "Perplexity Grade"}, "memory_exercise_words": {"items": {"type": "string"}, "type": "array", "title": "Memory Exercise Words"}, "memory_exercise_reply": {"type": "string", "title": "Memory Exercise Reply"}, "memory_grade": {"type": "object", "title": "Memory Grade", "default": {}}, "memory_exercise_reply_2": {"type": "string", "title": "Memory Exercise Reply 2"}, "memory_grade_2": {"type": "object", "title": "Memory Grade 2", "default": {}}, "l_reply": {"type": "string", "title": "L Reply"}, "l_grade": {"type": "object", "title": "L Grade", "default": {}}, "animal_reply": {"type": "string", "title": "Animal Reply"}, "animal_grade": {"type": "object", "title": "Animal Grade", "default": {}}, "dysarthria_grade": {"type": "number", "title": "Dysarthria Grade"}, "miscellaneous": {"type": "object", "title": "Miscellaneous", "default": {}}}, "type": "object", "required": ["call_sid", "phone_number", "timestamp", "previous_rejects", "history"], "title": "CallLog", "description": "CallLog(*, call_sid: str, phone_number: str, timestamp: datetime.datetime, participant_study_id: str = 'unknown', rejected: Optional[str] = 'false', previous_rejects: int, history: list[tuple[str, str]], syntax_grade: Optional[list] = [], perplexity_grade: Optional[float] = None, memory_exercise_words: Optional[List[str]] = None, memory_exercise_reply: Optional[str] = None, memory_grade: dict = {}, memory_exercise_reply_2: Optional[str] = None, memory_grade_2: dict = {}, l_reply: Optional[str] = None, l_grade: dict = {}, animal_reply: Optional[str] = None, animal_grade: dict = {}, dysarthria_grade: Optional[float] = None, miscellaneous: dict = {})"}, "CallLogHeader": {"properties": {"call_sid": {"type": "string", "title": "Call Sid"}, "participant_study_id": {"type": "string", "title": "Participant Study Id"}, "timestamp": {"type": "string", "format": "date-time", "title": "Timestamp"}}, "type": "object", "required": ["call_sid", "participant_study_id", "timestamp"], "title": "CallLogHeader", "description": "CallLogHeader(*, call_sid: str, participant_study_id: str, timestamp: datetime.datetime)"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Participant": {"properties": {"participant_study_id": {"type": "string", "title": "Participant Study Id"}, "phone_number": {"type": "string", "title": "Phone Number"}, "end_date": {"type": "string", "format": "date-time", "title": "End Date"}, "start_date": {"type": "string", "format": "date-time", "title": "Start Date"}, "morning_time": {"type": "integer", "title": "Morning Time"}, "afternoon_time": {"type": "integer", "title": "Afternoon Time"}, "evening_time": {"type": "integer", "title": "Evening Time"}, "morning_minute": {"type": "integer", "title": "Morning Minute"}, "afternoon_minute": {"type": "integer", "title": "Afternoon Minute"}, "evening_minute": {"type": "integer", "title": "Evening Minute"}}, "type": "object", "required": ["participant_study_id", "phone_number", "end_date", "start_date", "morning_time", "afternoon_time", "evening_time", "morning_minute", "afternoon_minute", "evening_minute"], "title": "Participant", "description": "Participant(*, participant_study_id: str, phone_number: str, end_date: datetime.datetime, start_date: datetime.datetime, morning_time: int, afternoon_time: int, evening_time: int, morning_minute: int, afternoon_minute: int, evening_minute: int)"}, "ScheduledCall": {"properties": {"id": {"type": "string", "title": "Id"}, "phone_number": {"type": "string", "title": "Phone Number"}, "time": {"type": "string", "format": "date-time", "title": "Time"}, "rejects": {"type": "integer", "title": "Rejects"}}, "type": "object", "required": ["id", "phone_number", "time", "rejects"], "title": "ScheduledCall", "description": "ScheduledCall(*, id: str, phone_number: str, time: datetime.datetime, rejects: int)"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}